% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fixed.r
\name{fit_fixed}
\alias{fit_fixed}
\title{fit_fixed() fit model with fixed model parameters to retrieve model realizations l (i.e., fixed calibration)}
\usage{
fit_fixed(data, scale_num, scale_info, model_param, file_name)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}

\item{model_param}{Mplus code for the model retrieved by get_response_model()}

\item{file_name}{Mplus file names}
}
\value{
mplus_object generated with MplusAutomation
}
\description{
fit_fixed() fit model with fixed model parameters to retrieve model realizations l (i.e., fixed calibration)
}
\examples{

mplus_out_eta  <- rd3c3::silent(rd3c3::fit_fixed(
data            = data_responses,
scale_num       = scale_id,
scale_info      = scales_data,
model_param     = get_response_model(grm_model),
file_name       = paste0(scale_file, '_realizations')
))

}
