% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pcm.r
\name{fit_pcm}
\alias{fit_pcm}
\title{fit_pcm() it fits a partial credit model (PCM) using MPLUS and MplusAutomation}
\usage{
fit_pcm(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
mplus_object generated with MplusAutomation
}
\description{
fit_pcm() it fits a partial credit model (PCM) using MPLUS and MplusAutomation
}
\examples{

scale_001 <- fit_pcm(
scale_info = scales_data,
scale_num  = 1,
data = data_model)


}
