% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reg.r
\name{fit_reg}
\alias{fit_reg}
\title{fit_reg() fits a linear regression model with pseudo maximum likelihood, using MPLUS and MplusAutomation, with and without controling by a covariate.
          in particular, it fits a population average model.}
\usage{
fit_reg(data, pv1, pv2, pv3, pv4, pv5, ind)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{pv1}{vector within the data frame with plausible value 1}

\item{pv2}{vector within the data frame with plausible value 2}

\item{pv3}{vector within the data frame with plausible value 3}

\item{pv4}{vector within the data frame with plausible value 4}

\item{pv5}{vector within the data frame with plausible value 5}

\item{ind}{vector name of the target variable of interest}
}
\value{
table with estimates
}
\description{
fit_reg() fits a linear regression model with pseudo maximum likelihood, using MPLUS and MplusAutomation, with and without controling by a covariate.
          in particular, it fits a population average model.
}
\examples{

est_table <- fit_mlm(
             data = data_response,
             pv1  = PV1CIV,
             pv2  = PV2CIV,
             pv3  = PV3CIV,
             pv4  = PV4CIV,
             pv5  = PV5CIV,
             ind  = opd
             )

}
