% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_realizations.r
\name{get_realizations}
\alias{get_realizations}
\title{get_realizations() produces realizations given a Mplus code model (i.e., fixed calibration)}
\usage{
get_realizations(data, scale_num, scale_info, model_param, file_name)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}

\item{model_param}{Mplus code for the model retrieved by get_response_model()}

\item{file_name}{Mplus file names}
}
\value{
mplus_object generated with MplusAutomation
}
\description{
get_realizations() produces realizations given a Mplus code model (i.e., fixed calibration)
}
\examples{

mplus_out_eta  <- rd3c3::silent(rd3c3::get_realizations(
data            = data_responses,
scale_num       = scale_id,
scale_info      = scales_data,
model_param     = get_response_model(grm_model),
file_name       = paste0(scale_file, '_realizations')
))

}
