% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_model.r
\name{get_response_model}
\alias{get_response_model}
\title{get_response_model() retrieves the mplus code to generate realizations}
\usage{
get_response_model(out)
}
\arguments{
\item{out}{it uses as an input a mplus_object generated with MplusAutomation}
}
\value{
a formula string with the mplus code to produce the scores
}
\description{
get_response_model() retrieves the mplus code to generate realizations
}
\examples{

get_response_model(grm_model)


}
