% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_caterpillar.r
\name{grm_caterpillar}
\alias{grm_caterpillar}
\title{grm_caterpillar() produces a caterpillar plot of theta realizations}
\usage{
grm_caterpillar(model)
}
\arguments{
\item{model}{it uses an input a mplus_object generated with MplusAutomation}
}
\value{
a numeric object
}
\description{
grm_caterpillar() produces a caterpillar plot of theta realizations
}
\examples{

grm_caterpillar(scale_001)


}
