% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_mrfa_full.r
\name{pa_mrfa_full}
\alias{pa_mrfa_full}
\title{pa_mrfa_full() calculates a parallel analysis using a response data frame}
\usage{
pa_mrfa_full(data, scale_num, scale_info, n_sim)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}

\item{n_sim}{specify the number of simulated correlations, n_sim = 500 is recommended}
}
\value{
a data frame generated with DA.MRFA::parallelMRFA
}
\description{
pa_mrfa_full() calculates a parallel analysis using a response data frame
}
\examples{

pa_mrfa(
data = scale_data,
scale_num = 1,
scale_info = items_data,
n_sim = 500
) \%>\%
knitr::kabel(., digits = 2)


}
