% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_plot.r
\name{pa_plot}
\alias{pa_plot}
\title{pa_plot() plot the results of a parallel analysis generated with the pa_mrfa() function}
\usage{
pa_plot(data)
}
\arguments{
\item{data}{it uses an input a pa_mrfa() results, generated with DA.MRFA::parallelMRFA}
}
\value{
a ggplot object
}
\description{
pa_plot() plot the results of a parallel analysis generated with the pa_mrfa() function
}
\examples{

plot_error(pa_mrfa_data)


}
