% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_fit_model.r
\name{re_fit_model}
\alias{re_fit_model}
\title{re_fit_model() produces realizations given a Mplus code model (i.e., fixed calibration)}
\usage{
re_fit_model(data, scale_num, scale_info, model_param, file_name)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}

\item{model_param}{Mplus code for the model retrieved by get_response_model()}

\item{file_name}{Mplus file names}
}
\value{
mplus_object generated with MplusAutomation
}
\description{
re_fit_model() produces realizations given a Mplus code model (i.e., fixed calibration)
}
\examples{

scale_001 <- fit_grm2(
scale_info = scales_data,
scale_num  = 1,
data = data_model)


}
