% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_empirical.r
\name{reliability_empirical}
\alias{reliability_empirical}
\title{reliability_empirical() calculates empirical reliability of theta scores realizations}
\usage{
reliability_empirical(model)
}
\arguments{
\item{model}{it uses an input a mplus_object generated with MplusAutomation}
}
\value{
a numeric object
}
\description{
reliability_empirical() calculates empirical reliability of theta scores realizations
}
\examples{

reliability_empirical(scale_001)


}
