% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_eta.R
\name{std_eta}
\alias{std_eta}
\title{std_eta() scales a theta score, or any other score into a standardized score, relying on the the survey design of the data}
\usage{
std_eta(data, eta)
}
\arguments{
\item{data}{a data frame, containing id_j, ws, and id_s vectors for survey design}

\item{eta}{a IRT or LVM realization (or any other score)}
}
\value{
a numeric vector with  with mean == 0 and sd == 1, at the population level
}
\description{
std_eta() scales a theta score, or any other score into a standardized score, relying on the the survey design of the data
}
\examples{

score_01 <- data_response \%>\%
            mutate(std_bel = std_eta(., eta_score))


}
