% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tam_caterpillar.r
\name{tam_caterpillar}
\alias{tam_caterpillar}
\title{tam_caterpillar() fits a rash models with TAM and generates a caterpillar plot with its theta realizations}
\usage{
tam_caterpillar(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
a caterpillar plot generated with library(ggplot2)
}
\description{
tam_caterpillar() fits a rash models with TAM and generates a caterpillar plot with its theta realizations
}
\examples{

missing_summary(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
knitr::kabel(., digits = 2)


}
